requirements.txt
# FastAPI core
fastapi==0.104.1
uvicorn[standard]==0.24.0


# data init and serialize
pydantic==2.5.0


# auth
PyJWT==2.8.0
python-multipart==0.0.6

# dev tools
pytest==7.4.3
httpx==0.25.2
requests==2.31.0

